2 commits
- clean unsued code
- orgainze temp
- current working on try- catch method for no search
- next -> display template for meal by card boostrap
