+++ grails-app/domain/MealStore/Person.groovy
@@ -1,0 +1,13 @@
+package MealStore
+
+class Person {
+        String name
+        Integer age
+        Date lastVisit
+        static hasMany  = [ingredients: Ingredient, measures: Measure]
+
+
+
+    static constraints = {
+    }
+}
