+++ grails-app/services/MealStore/MeasureService.groovy
@@ -1,4 +1,4 @@
-package meal1
+package MealStore
 
 import grails.gorm.services.Service
