+++ src/test/groovy/MealStore/PersonControllerSpec.groovy
@@ -1,9 +1,9 @@
-package newstore
+package MealStore
 
 import grails.testing.web.controllers.ControllerUnitTest
 import spock.lang.Specification
 
-class SearchControllerSpec extends Specification implements ControllerUnitTest<SearchController> {
+class PersonControllerSpec extends Specification implements ControllerUnitTest<PersonController> {
 
     def setup() {
     }
