--- src/test/groovy/meal1/IngredientSpec.groovy
@@ -1,18 +1,0 @@
-package meal1
-
-import grails.testing.gorm.DomainUnitTest
-import spock.lang.Specification
-
-class IngredientSpec extends Specification implements DomainUnitTest<Ingredient> {
-
-    def setup() {
-    }
-
-    def cleanup() {
-    }
-
-    void "test something"() {
-        expect:"fix me"
-            true == false
-    }
-}
