--- src/test/groovy/meal1/MealSpec.groovy
@@ -1,18 +1,0 @@
-package meal1
-
-import grails.testing.gorm.DomainUnitTest
-import spock.lang.Specification
-
-class MealSpec extends Specification implements DomainUnitTest<Meal> {
-
-    def setup() {
-    }
-
-    def cleanup() {
-    }
-
-    void "test something"() {
-        expect:"fix me"
-            true == false
-    }
-}
