--- src/test/groovy/newstore/SearchByNameServiceSpec.groovy
@@ -1,18 +1,0 @@
-package newstore
-
-import grails.testing.services.ServiceUnitTest
-import spock.lang.Specification
-
-class SearchByNameServiceSpec extends Specification implements ServiceUnitTest<SearchByNameService>{
-
-    def setup() {
-    }
-
-    def cleanup() {
-    }
-
-    void "test something"() {
-        expect:"fix me"
-            true == false
-    }
-}
