--- src/test/groovy/newstore/SearchByletterServiceSpec.groovy
@@ -1,18 +1,0 @@
-package newstore
-
-import grails.testing.services.ServiceUnitTest
-import spock.lang.Specification
-
-class SearchByletterServiceSpec extends Specification implements ServiceUnitTest<letter>{
-
-    def setup() {
-    }
-
-    def cleanup() {
-    }
-
-    void "test something"() {
-        expect:"fix me"
-            true == false
-    }
-}
